refactor: move useDrawerActions hook to hooks directory
- Extract useDrawerActions from drawer-options.tsx to src/hooks/use-drawer-actions.ts
- Update navigation index to re-export hook from new location
- Maintain backward compatibility for existing imports via @/components/layout
- Remove unused imports from drawer-options.tsx
